refactor: split preview send feature
